Changes for NUnit 2.5 (removed NUnit.Framework.SyntaxHelpers)
There are some minor changes in 'using' in tests all over the project
to make it compatible with NUnit 2.5.

Advantages of working with NUnit 2.5+:  [TestCase(...)] attribute
allows parameterized tests; new parameters in [TextFixture] allow
generic and/or parameterized "fixtures" (test classes)

IMPORTANT: if you apply this patch, make sure (a) to install
the most recent NUnit and (b) to change it in References from
NUnit.Framework 2.4x to 2.5.5, for ALL assemblies!