Tests for the new Migration unit (requires NUnit 2.5 or better !)
The new Migration unit is fully tested for MySQL.  IMPORTANT: after adding these new
files to the Resources directory, make sure to set "Build Action"
to "Embedded resource" in the properties (in VS, or whatever is the
equivalent on Mono)

Tests for other DBs can be added
by copying the [TextFixture] attribute and modifying its parameters
accordingly. The "MigrProc_MySQL.migrations" test file should be copied
to "MigrProc_MSSQL.migrations" and the functions inside modified accordingly.

It should be possible to use the Migration class directly
for databases other than MySQL (the MySqlMigration class is a workaround
for MySQL'a weird behaviour when defining procs/funcs).