Added GetMetadata() and Delete() methods to IAssetData,
AssertService modified accordingly to use it. The "Delete()" is
for completeness, block it somewhere (in AssetService?) if its not
really needed.  The "delete asset" console command is also implemented.

"GetMetadata()" was emulated in AssetService by getting
the entire asset and dropping the data. It is more efficient
to handle this on DB-connector level and NOT read asset data from the
DB if only metadata is needed.

NO DBMS-implementations of the new methods at this point, yet.