(minor correction, doesn't affect anything)
Most of the useful stuff from MSSQLManager.cs has been replicated in
either new DBMS-independent base classes or in MSSQLDataSpecific.cs,
so if the new class structure will be accepted (and other MS SQL data
modules refactored accordingly), this file won't be needed.

Meanwhile, the file is still in use so I didn't do anything bad to it.