AssetStore.migrations is back there
NOTE: the last 2 sections of this migration file are currently
disabled. If enabled (see comments), they would convert the asset
table to use BINARY(16) instead of CHAR(36) for both ID and CreatorID.

NOTE: Changing the data type in the table does NOT require any changes
in the (new) code!!!  You can easily run the performance tests with either
version of the table.

This is supposed to make indexing faster, but so far I couldn't see
any effect of that.