Added SQLite conn string in TestDataConnection.ini
Although SQLite tests can be run with empty conn string
(using a temporary file for the DB), the performance
tests need the DB in a permanent file.

Not sure where the DB file will end up if specified
without a path (as it's done here), change it as suitable
for your system.