List of types in BaseDataBase.Cmd .ctor is made optional
In most cases each '@param' in SQL query corresponds to a
table column and their names match (e.g. 'WHERE id=@id').
We fetch the column info from the DB and map it to parameters,
so no type info for parameters usually need to be specified.
We still need to know the types for params like those in
'LIMIT @first, @count'

Signed-off-by: AlexRa <info@adeptsql.com>